// src/generated-types/menu-category.ts
export interface MenuCategoryTranslation {
  id: string;
  language: string;
  title: string;
  description?: string | null | undefined;
}
// ...
export interface CreateMenuCategoryRequest {
  slug: string;
  imageUrl?: string | null | undefined;
  isAvailable?: boolean | null | undefined;
}

/** Request message for updating a menu category. */
export interface UpdateMenuCategoryRequest {
  id: string;
  slug?: string | null | undefined;
  imageUrl?: string | null | undefined;
  isAvailable?: boolean | null | undefined;
}

/** Request message for changing the position of a menu category. */
export interface ChangeMenuCategoryPositionRequest {
  id: string;
  position: number;
}

/** Request message for creating a menu category translation. */
export interface CreateMenuCategoryTranslationRequest {
  categoryId: string;
  language: string;
  title: string;
  description?: string | null | undefined;
}

/** Request message for updating a menu category translation. */
export interface UpdateMenuCategoryTranslationRequest {
  id: string;
  title?: string | null | undefined;
  description?: string | null | undefined;
}